Move reusable helper methods out of margin mixin
