+++ src/mixins/margin.js
@@ -1,10 +1,4 @@
-const getUnit = value => (value === 0 ? "0" : `${value}px`)
-
-export const getDimension = (theme, size) =>
-  typeof size === "number" ? getUnit(theme.constants.SIZE_SUB_UNIT * size) : "auto"
-
-export const getDimensions = (theme, value) =>
-  value.map(size => getDimension(theme, size)).join(" ")
+import { getDimensions } from "src/mixins/utils"
 
 export default ({ theme, margin }) => {
   if (!margin) return ""
